chore: consolidate test archives and rename ~~normalizer.rs~~ module to normalize.rs
